Rearrange the APNS module's code in a modular structure. No functionality changes.
