+++ srv/apns/payload.go
@@ -1,0 +1,102 @@
+/*
+ * Copyright 2011-2013 Nan Deng
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *	http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package apns
+
+// Contains functions for building a payload from the url parameter abstraction.
+
+import (
+	"strconv"
+
+	"github.com/uniqush/uniqush-push/push"
+	"github.com/uniqush/uniqush-push/srv/apns/common"
+)
+
+func toAPNSPayload(n *push.Notification) ([]byte, push.PushError) {
+	payload := make(map[string]interface{})
+	aps := make(map[string]interface{})
+	alert := make(map[string]interface{})
+	for k, v := range n.Data {
+		switch k {
+		case "msg":
+			alert["body"] = v
+		case "action-loc-key":
+			alert[k] = v
+		case "loc-key":
+			alert[k] = v
+		case "loc-args":
+			alert[k] = parseList(v)
+		case "badge":
+			b, err := strconv.Atoi(v)
+			if err != nil {
+				continue
+			} else {
+				aps["badge"] = b
+			}
+		case "sound":
+			aps["sound"] = v
+		case "content-available":
+			aps["content-available"] = v
+		case "img":
+			alert["launch-image"] = v
+		case "id":
+			continue
+		case "expiry":
+			continue
+		case "ttl":
+			continue
+		default:
+			payload[k] = v
+		}
+	}
+
+	aps["alert"] = alert
+	payload["aps"] = aps
+	j, err := common.MarshalJSONUnescaped(payload)
+	if err != nil {
+		return nil, push.NewErrorf("Failed to convert notification data to JSON: %v", err)
+	}
+	if len(j) > maxPayLoadSize {
+		return nil, push.NewBadNotificationWithDetails("payload is too large")
+	}
+	return j, nil
+}
+
+func parseList(str string) []string {
+	ret := make([]string, 0, 10)
+	elem := make([]rune, 0, len(str))
+	escape := false
+	for _, r := range str {
+		if escape {
+			escape = false
+			elem = append(elem, r)
+		} else if r == '\\' {
+			escape = true
+		} else if r == ',' {
+			if len(elem) > 0 {
+				ret = append(ret, string(elem))
+			}
+			elem = elem[:0]
+		} else {
+			elem = append(elem, r)
+		}
+	}
+	if len(elem) > 0 {
+		ret = append(ret, string(elem))
+	}
+	return ret
+}
